requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.15.0=pypi_0
addict=2.4.0=pypi_0
anyio=3.5.0=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
attrs=21.4.0=pypi_0
autopep8=1.6.0=pyhd3eb1b0_1
babel=2.10.1=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
bleach=5.0.0=pypi_0
bottleneck=1.3.2=py38heb32a55_1
brotli=1.0.9=he6710b0_2
ca-certificates=2022.07.19=h06a4308_0
cachetools=5.0.0=pypi_0
carla=0.9.9=dev_0
certifi=2022.6.15=py38h06a4308_0
cffi=1.15.0=pypi_0
charset-normalizer=2.0.12=pypi_0
clang=5.0=pypi_0
cloudpickle=2.0.0=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cvxpy=1.2.0=pypi_0
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.30=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pypi_0
deprecation=2.1.0=pypi_0
ecos=2.0.10=pypi_0
elementpath=2.5.3=pypi_0
entrypoints=0.3=py38_0
ephem=4.1.2=py38h7f8727e_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.4=h295c915_0
fasteners=0.17.3=pypi_0
fastjsonschema=2.15.3=pypi_0
flatbuffers=1.12=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
gast=0.4.0=pypi_0
giflib=5.2.1=h7b6447c_0
glfw=2.5.3=pypi_0
glib=2.69.1=h4ff587b_1
google-auth=2.6.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.44.0=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
gym=0.23.1=pypi_0
gym-algorithmic=0.0.1=pypi_0
gym-notices=0.0.6=pypi_0
h5py=3.1.0=pypi_0
icu=58.2=he6710b0_3
idna=3.3=pypi_0
imageio=2.19.2=pypi_0
importlib-metadata=4.11.3=pypi_0
importlib-resources=5.7.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=py38h06a4308_0
ipython=8.1.1=py38h06a4308_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.0=pypi_0
jedi=0.18.1=py38h06a4308_1
jinja2=3.1.1=pypi_0
joblib=1.1.0=pypi_0
jpeg=9d=h7f8727e_0
json5=0.9.6=pypi_0
jsonschema=4.4.0=pypi_0
jupyter-packaging=0.12.0=pypi_0
jupyter-server=1.16.0=pypi_0
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_core=4.9.2=py38h06a4308_0
jupyterlab=3.3.4=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.13.0=pypi_0
jupyterlab-widgets=1.1.0=pypi_0
keras=2.6.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.2=py38h295c915_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.6=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=py38h06a4308_1
matplotlib-base=3.5.1=py38ha18d171_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mujoco-py=2.0.2.13=pypi_0
munkres=1.1.4=py_0
nbclassic=0.3.7=pypi_0
nbclient=0.6.0=pypi_0
nbconvert=6.5.0=pypi_0
nbformat=5.3.0=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.8.4=pypi_0
notebook=6.4.11=pypi_0
notebook-shim=0.1.0=pypi_0
numexpr=2.8.1=py38h6abb31d_0
numpy=1.19.5=pypi_0
numpy-base=1.22.3=py38hf524024_0
oauthlib=3.2.0=pypi_0
open3d=0.15.2=pypi_0
openssl=1.1.1q=h7f8727e_0
opt-einsum=3.3.0=pypi_0
osqp=0.6.2.post5=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=py38h295c915_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py38h22f2fdc_0
pip=21.2.4=py38h06a4308_0
prometheus-client=0.14.1=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.4=pypi_0
psutil=5.8.0=py38h27cfd23_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py-trees=0.8.3=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycodestyle=2.8.0=pyhd3eb1b0_0
pycparser=2.21=pypi_0
pydot=1.4.2=pypi_0
pygame=2.1.2=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pyquaternion=0.9.9=pypi_0
pyrsistent=0.18.1=pypi_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=py38h295c915_2
qdldl=0.1.5.post2=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
scs=3.2.0=pypi_0
seaborn=0.12.0=pypi_0
send2trash=1.8.0=pypi_0
setuptools=62.1.0=pypi_0
shapely=1.8.2=pypi_0
simple-watchdog-timer=0.1.1=pypi_0
sip=4.19.13=py38h295c915_0
six=1.15.0=pypi_0
sklearn=0.0=pypi_0
sniffio=1.2.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.38.0=hc218d9a_0
stack_data=0.2.0=pyhd3eb1b0_0ß
tabulate=0.8.10=py38h06a4308_0
# tensorboard=2.8.0=pypi_0
# tensorboard-data-server=0.6.1=pypi_0
# tensorboard-plugin-wit=1.8.1=pypi_0
# tensorflow-estimator=2.6.0=pypi_0
# tensorflow-gpu=2.6.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.13.3=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.1.1=pypi_0
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
tomlkit=0.10.2=pypi_0
tornado=6.1=py38h27cfd23_0
tqdm=4.64.0=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.26.9=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=pypi_0
websocket-client=1.3.2=pypi_0
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.6.0=pypi_0
wrapt=1.12.1=pypi_0
xmlschema=1.11.3=pypi_0
xz=5.2.5=h7b6447c_0
yapf=0.31.0=pyhd3eb1b0_0
zeromq=4.3.4=h2531618_0
zipp=3.7.0=pypi_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
